[fix] embedded.mobileprovision file problem

+++ Tools/pack.sh
@@ -95,8 +95,10 @@ function pack(){
 		unzip -oqq "${TARGET_IPA_PATH}" -d "${TEMP_PATH}"
 		cp -rf "${TEMP_PATH}/Payload/"*.app "${TARGET_APP_PUT_PATH}"
 	fi
-
-	mv "${BUILD_APP_PATH}/embedded.mobileprovision" "${BUILD_APP_PATH}"/..
+	
+	if [ -f "${BUILD_APP_PATH}/embedded.mobileprovision" ]; then
+		mv "${BUILD_APP_PATH}/embedded.mobileprovision" "${BUILD_APP_PATH}"/..
+	fi
 
 	#remove origin .app
 	rm -rf "${BUILD_APP_PATH}" || true
@@ -115,7 +117,9 @@ function pack(){
 		cp -rf "${COPY_APP_PATH}/" "${BUILD_APP_PATH}/"
 	fi
 
-	mv "${BUILD_APP_PATH}/../embedded.mobileprovision" "${BUILD_APP_PATH}"
+	if [ -f "${BUILD_APP_PATH}/../embedded.mobileprovision" ]; then
+		mv "${BUILD_APP_PATH}/../embedded.mobileprovision" "${BUILD_APP_PATH}"
+	fi
 
 	# get target info
 	ORIGIN_BUNDLE_ID=$(/usr/libexec/PlistBuddy -c "Print CFBundleIdentifier"  "${COPY_APP_PATH}/Info.plist" 2>/dev/null)
